requirements.txt
discord.py
pymongo
umongo
dnspython